implementacao de funcaoes para criar datas inicio e fim para busca
