+++ src/model/Despesas.js
@@ -42,7 +42,7 @@ const despesas = new mongoose.Schema(
   },
   {
     versionKey: false,
-    timestamps: true,
+    timestamps: false,
   }
 );
